Don't pass in RAILS ENV in dev docker-compose.yml
Because by default, rails runs as development environment. And the problem,
with this is if we pass in 'RAILS_ENV', running tests with this command:
docker-compose exec api bundle exec rspec
It still runs the test in 'development' environment. See 'rails_helper.rb' for details
on how RAILS_ENV is setup during test runs.

[#128067971]