Index Person documents in ES when a person is created
One new folders are created in 'app/' called 'search_repos' which houses the
Repository objects for various indices in ES. Both the PersonIndexer and the
Search Repositories are in the API service at the moment. If we want a different
service responsible for indexing the documents, we will need to  move files from
these two folders.

[#128067971]